refactor(play): split PlayerCharacterCreateView into composable + two focused components
Extract all form state and logic into useCharacterCreationForm (composable),
shared via provide/inject. CharacterCreateWizard owns the 8-step creation
wizard template; CharacterEditTabs owns the 3-tab edit form. The view is now
a 12-line shell. Each file is independently readable and indexable.
