+++ ROADMAP.md
@@ -527,3 +527,5 @@ Keep the core DM tooling free forever (open source). Gate AI features, advanced 
 - [x] **Classes: DB-driven caster_type** (irongollem/grimoire#152) — added `caster_type`, `prepared_ability`, `prepared_divisor` columns to `system_classes` and `custom_classes`; seeded all 13 SRD classes with correct values; `useClassByName()` composable returns system-or-custom class by name; replaced `getCasterType`/`getMaxPrepared`/`getSlotRecovery`/`getHitDie` static calls in PlayerSpellsView, PlayerCharacterView, RestButtons, RestDialog, PlayerFeaturesTab, RunnerEntityDetail with DB-driven fallbacks; CustomClassEditor gains caster_type + prepared_ability + prepared_divisor UI in Spellcasting section
 
 - [x] **Classes: Delete static TS class files** (irongollem/grimoire#153) — removed all 13 `src/levelup/classes/*.ts` files and `classFeatures.ts`; LevelUpWizard now derives ASI levels, subclass level, spell slots, resources, and steps from `systemClass`/`customClass` DB data; `PARTY_CLASSES`/`PartyClass` removed from `party.types.ts`; PlayerCharacterCreateView and CustomSubclassEditorView now build class name lists from DB; PlayerFeaturesTab now resolves feature UUIDs via `useAllFeatures` instead of static table
+
+- [x] **Encounter runner: ranged attacks consume ammunition** (irongollem/grimoire#158) — "Ranged Attacks" section added below Melee Attacks in `RunnerEntityDetail.vue` for player combatants; shows any weapon equipped in main/off-hand that has the `"ammunition"` property (bows, crossbows, slings, blowguns, firearms, etc.); weapon→ammo tag mapping (arrow/bolt/bullet/needle/dart/firearm-bullet) resolves via vault item subtype/tags with name-pattern fallback; on attack roll, one unit of ammo is consumed — charge-tracked packs (e.g. Arrows (20)) decrement `current_charges`, single ammo decrements `quantity` and auto-removes at 0; attack button disabled with "— no ammo" label when no compatible ammo found in player's inventory containers/belt/backpack; ammo count badge shows remaining shots
